Experimental "filenames_cache_dir" option implementation
This option allows to speed-up repeatitive runs of application which was
not changed between runs.
When cache directory is set an application would try to find source file
names inside it when any module is loaded and use them for further
analysis instead of requesting all source file names and filter them again.

* Increase toolchain version to v143
* Add additional nuget dependencies (nlohmann json and leveldb)
* Extract IDebugInformationHandler to standalone file
* Introduce IModuleSourceFilenamesCache interface
* Introduce LevelDbModuleSourceFilenamesCache implementation using leveldb
  library
* Introduce DummyModuleSourceFilenamesCache for running without cache